Criando e configurando models Usuario e Produto

+++ script.sql
@@ -12,5 +12,5 @@ create table Produtos(
 Id int primary key auto_increment,
 Nome varchar(50) not null,
 Descricao varchar(200) not null,
-Preco float not null
+Preco decimal(5,2) not null
 );
